Remove Go <1.18 compile support for version VCS IDs
We are far beyond requiring a newer Go version than this.

No functional change.
